server/requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.52.0
django-cors-headers==3.14.0
django-filter==23.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Markdown==3.4.3
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.15
zope.interface==6.0
